fixed #13 迁移至 MD3

+++ app/src/main/java/com/equationl/giteetodo/ui/page/RepoList.kt
@@ -14,20 +14,25 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Card
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Snackbar
-import androidx.compose.material.SnackbarHost
-import androidx.compose.material.SnackbarResult
-import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.outlined.LibraryAdd
-import androidx.compose.material.rememberScaffoldState
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.DismissDirection
+import androidx.compose.material3.DismissValue
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Snackbar
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarResult
+import androidx.compose.material3.SwipeToDismiss
+import androidx.compose.material3.Text
+import androidx.compose.material3.rememberBottomSheetScaffoldState
+import androidx.compose.material3.rememberDismissState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.rememberCoroutineScope
@@ -51,12 +56,10 @@ import androidx.paging.compose.itemsIndexed
 import com.equationl.giteetodo.R
 import com.equationl.giteetodo.data.user.model.response.Repo
 import com.equationl.giteetodo.ui.common.Route
-import com.equationl.giteetodo.ui.theme.baseBackground
 import com.equationl.giteetodo.ui.widgets.BlurImage
 import com.equationl.giteetodo.ui.widgets.LinkText
 import com.equationl.giteetodo.ui.widgets.ListEmptyContent
 import com.equationl.giteetodo.ui.widgets.LoadDataContent
-import com.equationl.giteetodo.ui.widgets.SwipeableActionCard
 import com.equationl.giteetodo.ui.widgets.TopBar
 import com.equationl.giteetodo.util.Utils
 import com.equationl.giteetodo.viewmodel.RepoListViewAction
@@ -67,6 +70,7 @@ import kotlinx.coroutines.launch
 
 private const val TAG = "el, RepoList"
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun RepoListScreen(
     navController: NavHostController,
@@ -74,7 +78,7 @@ fun RepoListScreen(
 ) {
     val activity = (LocalContext.current as? Activity)
     val viewState = viewModel.viewStates
-    val scaffoldState = rememberScaffoldState()
+    val scaffoldState = rememberBottomSheetScaffoldState()
     val coroutineState = rememberCoroutineScope()
     val uriHandler = LocalUriHandler.current
 
@@ -86,13 +90,18 @@ fun RepoListScreen(
                         scaffoldState.snackbarHostState.showSnackbar(message = it.message)
                     }
                 }
+
                 is RepoListViewEvent.Goto -> {
                     navController.navigate(it.route)
                 }
+
                 is RepoListViewEvent.ShowDeleteRepoMsg -> {
                     coroutineState.launch {
-                        when (scaffoldState.snackbarHostState.showSnackbar(message = it.message, actionLabel = "前往删除")) {
-                            SnackbarResult.Dismissed -> {  }
+                        when (scaffoldState.snackbarHostState.showSnackbar(
+                            message = it.message,
+                            actionLabel = "前往删除"
+                        )) {
+                            SnackbarResult.Dismissed -> {}
                             SnackbarResult.ActionPerformed -> {
                                 uriHandler.openUri(it.deleteUrl)
                             }
@@ -118,8 +127,7 @@ fun RepoListScreen(
                 val lastQueue = backStackList[backStackList.size - 2]
                 if (lastQueue.destination.route?.contains(Route.HOME) == true) { // 只有从首页跳转过来的才返回，否则直接退出
                     navController.popBackStack()
-                }
-                else {
+                } else {
                     activity?.finish()
                 }
             }
@@ -127,13 +135,14 @@ fun RepoListScreen(
         snackbarHost = {
             SnackbarHost(hostState = scaffoldState.snackbarHostState) { snackBarData ->
                 Snackbar(snackbarData = snackBarData)
-            }})
+            }
+        })
     {
         RepoListContent(viewState.repoFlow, viewModel, it)
     }
 }
 
-@OptIn(ExperimentalMaterialApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun RepoListContent(
     repoFlow: Flow<PagingData<Repo>>,
@@ -145,52 +154,70 @@ fun RepoListContent(
     Log.i(TAG, "RepoListContent: loadState=${repoList.loadState}")
 
     if (repoList.loadState.refresh is LoadState.Error) {
-        viewModel.dispatch(RepoListViewAction.SendMsg("加载错误："+ (repoList.loadState.refresh as LoadState.Error).error.message))
+        viewModel.dispatch(RepoListViewAction.SendMsg("加载错误：" + (repoList.loadState.refresh as LoadState.Error).error.message))
     }
 
     if (repoList.itemCount < 1) {
         if (repoList.loadState.refresh == LoadState.Loading) {
             LoadDataContent("正在加载仓库列表中…")
-        }
-        else {
+        } else {
             ListEmptyContent("没有找到仓库数据，点击刷新或点击右上角创建一个仓库") {
                 repoList.refresh()
             }
         }
-    }
-    else {
+    } else {
         Column(
             Modifier
                 .fillMaxSize()
                 .padding(paddingValues)
-                .background(MaterialTheme.colors.baseBackground)) {
+                .background(MaterialTheme.colorScheme.background)
+        ) {
             LazyColumn {
-                itemsIndexed(repoList, key = {_, item -> item.fullName}) { _, item ->
+                itemsIndexed(repoList, key = { _, item -> item.fullName }) { _, item ->
                     if (item != null && item.namespace.type == "personal") {  // 仅加载类型为个人的仓库
-                        SwipeableActionCard(
-                            mainCard = {
-                                RepoItem(item) {
-                                    viewModel.dispatch(RepoListViewAction.ChoiceARepo(it))
+                        val dismissState = rememberDismissState(
+                            confirmValueChange = {
+                                if (it == DismissValue.DismissedToStart) {
+                                    viewModel.dispatch(RepoListViewAction.DeleteRepo(item.fullName))
                                 }
+                                false
                             },
-                            leftSwipeCard = {
-                                Card(Modifier.padding(32.dp),
-                                    shape = RoundedCornerShape(16.dp), elevation = 5.dp) {
+                        )
+
+                        SwipeToDismiss(
+                            state = dismissState,
+                            background = {
+                                Card(
+                                    modifier = Modifier.padding(32.dp),
+                                    shape = RoundedCornerShape(16.dp),
+                                    elevation = CardDefaults.cardElevation(defaultElevation = 5.dp)
+                                ) {
                                     Row(
                                         Modifier
                                             .fillMaxWidth()
                                             .height(100.dp)
-                                            .background(MaterialTheme.colors.error),
+                                            .background(MaterialTheme.colorScheme.error),
                                         horizontalArrangement = Arrangement.End,
-                                        verticalAlignment = Alignment.CenterVertically) {
-                                        Text("继续滑动删除", color = MaterialTheme.colors.background)
-                                        Icon(Icons.Filled.Delete, contentDescription = "删除", tint = MaterialTheme.colors.background)
+                                        verticalAlignment = Alignment.CenterVertically
+                                    ) {
+                                        Text(
+                                            "继续滑动删除",
+                                            color = MaterialTheme.colorScheme.background
+                                        )
+                                        Icon(
+                                            Icons.Filled.Delete,
+                                            contentDescription = "删除",
+                                            tint = MaterialTheme.colorScheme.background
+                                        )
                                     }
                                 }
                             },
-                            leftSwiped = {
-                                viewModel.dispatch(RepoListViewAction.DeleteRepo(item.fullName))
-                            }
+                            dismissContent = {
+                                RepoItem(item) {
+                                    viewModel.dispatch(RepoListViewAction.ChoiceARepo(it))
+                                }
+                            },
+                            directions = setOf(DismissDirection.EndToStart)
                         )
                     }
                 }
@@ -199,16 +226,23 @@ fun RepoListContent(
                     when (repoList.loadState.append) {
                         is LoadState.NotLoading -> {}
                         LoadState.Loading -> {
-                            Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
+                            Row(
+                                Modifier.fillMaxWidth(),
+                                horizontalArrangement = Arrangement.Center
+                            ) {
                                 Text("加载中")
                             }
                         }
+
                         is LoadState.Error -> {
-                            Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
+                            Row(
+                                Modifier.fillMaxWidth(),
+                                horizontalArrangement = Arrangement.Center
+                            ) {
                                 LinkText("加载失败，点击重试") {
                                     repoList.retry()
                                 }
-                                viewModel.dispatch(RepoListViewAction.SendMsg("加载出错："+ (repoList.loadState.append as LoadState.Error).error.toString()))
+                                viewModel.dispatch(RepoListViewAction.SendMsg("加载出错：" + (repoList.loadState.append as LoadState.Error).error.toString()))
                             }
                         }
                     }
@@ -218,16 +252,19 @@ fun RepoListContent(
     }
 }
 
-@OptIn(ExperimentalMaterialApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun RepoItem(itemData: Repo, onClick: (path: String) -> Unit) {
-    Card(onClick = { onClick.invoke(itemData.fullName) },
-        modifier = Modifier.padding(32.dp), shape = RoundedCornerShape(16.dp), elevation = 5.dp) {
+    Card(
+        onClick = { onClick.invoke(itemData.fullName) },
+        modifier = Modifier.padding(32.dp), shape = RoundedCornerShape(16.dp), elevation = CardDefaults.cardElevation(defaultElevation = 5.dp)
+    ) {
         Column {
             Box(
                 Modifier
                     .fillMaxWidth()
-                    .height(100.dp)) {
+                    .height(100.dp)
+            ) {
                 BlurImage(
                     paint = painterResource(R.drawable.bg2),  // 如果使用随机的背景而非固定背景会造成卡顿与闪退（猜测是内存溢出）
                     contentDescription = "background",
@@ -235,13 +272,21 @@ fun RepoItem(itemData: Repo, onClick: (path: String) -> Unit) {
                     modifier = Modifier.fillMaxSize()
                 )
 
-                Column(modifier = Modifier
-                    .fillMaxSize()) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxSize()
+                ) {
                     Row(
                         Modifier
                             .fillMaxWidth()
-                            .padding(start = 4.dp), horizontalArrangement = Arrangement.Start) {
-                        Text(text = itemData.path, color = Color.White, maxLines = 1, overflow = TextOverflow.Ellipsis)
+                            .padding(start = 4.dp), horizontalArrangement = Arrangement.Start
+                    ) {
+                        Text(
+                            text = itemData.path,
+                            color = Color.White,
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis
+                        )
                     }
 
                     Row(
@@ -249,7 +294,8 @@ fun RepoItem(itemData: Repo, onClick: (path: String) -> Unit) {
                         horizontalArrangement = Arrangement.Center,
                         verticalAlignment = Alignment.CenterVertically
                     ) {
-                        val text = if (itemData.openIssuesCount > 0) "${itemData.openIssuesCount}项未完成" else "已全部完成"
+                        val text =
+                            if (itemData.openIssuesCount > 0) "${itemData.openIssuesCount}项未完成" else "已全部完成"
                         Text(text = text, color = Color.White)
                     }
                 }
@@ -258,14 +304,27 @@ fun RepoItem(itemData: Repo, onClick: (path: String) -> Unit) {
                 Row(
                     Modifier
                         .fillMaxWidth()
-                        .weight(1f), verticalAlignment = Alignment.CenterVertically, horizontalArrangement = Arrangement.Center) {
-                    Text(text = itemData.name, fontSize = 20.sp, fontWeight = FontWeight.Bold, maxLines = 1, overflow = TextOverflow.Ellipsis, modifier = Modifier.padding(start = 4.dp))
+                        .weight(1f),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.Center
+                ) {
+                    Text(
+                        text = itemData.name,
+                        fontSize = 20.sp,
+                        fontWeight = FontWeight.Bold,
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis,
+                        modifier = Modifier.padding(start = 4.dp)
+                    )
                 }
                 Row(
                     Modifier
                         .fillMaxWidth()
                         .weight(1f)
-                        .padding(top = 8.dp), verticalAlignment = Alignment.CenterVertically, horizontalArrangement = Arrangement.Center) {
+                        .padding(top = 8.dp),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.Center
+                ) {
                     Text(text = Utils.getDateTimeString(itemData.createdAt), fontSize = 10.sp)
                 }
             }
